+++ README.md
@@ -3,13 +3,6 @@
 ### 개요
 트위터 인플루언서의 트윗이 아마존 제품 판매에 미치는 영향 분석을 위한 데이터 파이프라인을 구축한 리포지토리입니다.
 
-### 실행방법
-```
-python .\connect_google_drive_sqlite_db.py
-python .\amazon_product\collect_executor.py
-```
-
-
 *reference*
 1. [트위터 API](https://rapidapi.com/omarmhaimdat/api/twitter154)
 2. [아마존 API](https://rapidapi.com/letscrape-6bRBa3QguO5/api/real-time-amazon-data)
