--- amazon_product/pyproject.toml
@@ -1,17 +1,0 @@
-[tool.poetry]
-name = "amazon-product"
-version = "0.1.0"
-description = ""
-authors = ["Haerin Kim <lini1634@naver.com>"]
-readme = "README.md"
-
-[tool.poetry.dependencies]
-python = "^3.12"
-requests = "^2.31.0"
-SQLAlchemy = "^2.0.30"
-beautifulsoup4 = "^4.12.3"
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
