--- twitter_user/pyproject.toml
@@ -1,16 +1,0 @@
-[tool.poetry]
-name = "twitter-user"
-version = "0.1.0"
-description = ""
-authors = ["Haerin Kim <lini1634@naver.com>"]
-readme = "README.md"
-
-[tool.poetry.dependencies]
-python = "^3.12"
-requests = "^2.31.0"
-SQLAlchemy = "^2.0.30"
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
