+++ 메모.txt
@@ -1,0 +1,14 @@
+merge
+
+하나의 프로그램은 하나의 venv...
+폴더를 구분하는 기준은 공통 모듈,..
+1. src 상위폴더 생성.....
+2. src 내부: 공통 모듈 common, utils등의 이름으로 생성...
+3. src 외부: 환경변수, 깃헙액션.....
+4. src 내부: models안에 테이블들...(db명세)
+** mvc(모델 뷰 컨트롤러) 패턴, mvvm, 모노레포?. 터보레포..(모놀리식<->마이크로)
+
+깃헙 액션으로 병렬 수집 가능..
+
+
+* 대시보드 기획-> 개발 (메타베이스)..
